+++ validx/py/chars.py
@@ -13,6 +13,10 @@ class Str(abstract.Validator):
     :param bool nullable:
         accept ``None`` as a valid value.
 
+    :param bool coerce:
+        convert non-string value to ``str``,
+        **use with caution** (see notes below).
+
     :param bool dontstrip:
         do not strip leading & trailing whitespace.
 
@@ -54,10 +58,19 @@ class Str(abstract.Validator):
     :raises OptionsError:
         if ``value not in self.options``.
 
+
+    :note:
+        Since any Python object can be converted to a string,
+        using ``coerce`` without other checks in fact validates nothing.
+        It can be useful though to sanitize data from sources with automatic type inferring,
+        where string data might be incorrectly interpreted as another type.
+        For example, phone number as ``int``, version number as ``float``, etc.
+
     """
 
     __slots__ = (
         "nullable",
+        "coerce",
         "dontstrip",
         "normspace",
         "encoding",
@@ -70,6 +83,7 @@ class Str(abstract.Validator):
     def __init__(
         self,
         nullable=False,
+        coerce=False,
         dontstrip=False,
         normspace=False,
         encoding=None,
@@ -81,6 +95,7 @@ class Str(abstract.Validator):
         replace=False,
     ):
         nullable = contracts.expect_flag(self, "nullable", nullable)
+        coerce = contracts.expect_flag(self, "coerce", coerce)
         dontstrip = contracts.expect_flag(self, "dontstrip", dontstrip)
         normspace = contracts.expect_flag(self, "normspace", normspace)
         encoding = contracts.expect_str(self, "encoding", encoding, nullable=True)
@@ -93,6 +108,7 @@ class Str(abstract.Validator):
 
         setattr = object.__setattr__
         setattr(self, "nullable", nullable)
+        setattr(self, "coerce", coerce)
         setattr(self, "dontstrip", dontstrip)
         setattr(self, "normspace", normspace)
         setattr(self, "encoding", encoding)
@@ -112,6 +128,8 @@ class Str(abstract.Validator):
                     value = value.decode(self.encoding)
                 except UnicodeDecodeError:
                     raise exc.StrDecodeError(expected=self.encoding, actual=value)
+            elif self.coerce:
+                value = str(value)
             else:
                 raise exc.InvalidTypeError(expected=str, actual=type(value))
         if not self.dontstrip: