+++ docs/changes.rst
@@ -18,6 +18,9 @@ Changes
 *   Added ability to normalize spaces by :class:`validx.py.Str` validator,
     i.e. replace space sequences by single space character.
     Use ``normspace=True`` parameter to enable the normalization.
+*   Added ``coerce`` parameter to :class:`validx.py.Str` validator,
+    which can be useful to sanitize data from sources with automatic type inferring,
+    where string data might be incorrectly interpreted as another type.
 *   Unified behavior of Python and Cython versions of :class:`validx.py.Int` validator.
     Both versions support unlimited integer numbers now.
 *   Added new :class:`validx.py.Set` validator.